add pagination to SearchCourseActivity.kt
