+++ app/build.gradle
@@ -140,6 +140,7 @@ dependencies {
     implementation "io.noties.markwon:linkify:$markwon_version"
     implementation 'com.airbnb.android:lottie:3.4.0'
     implementation 'com.github.bumptech.glide:glide:4.11.0'
+    implementation 'androidx.paging:paging-runtime-ktx:2.1.1'
     kapt 'com.github.bumptech.glide:compiler:4.11.0'
     implementation 'com.hbb20:ccp:2.4.0'
