+++ app/src/main/java/com/codingblocks/cbonlineapp/course/SearchCourseActivity.kt
@@ -4,23 +4,27 @@ import android.content.Intent
 import android.os.Bundle
 import android.widget.ImageView
 import androidx.core.app.ActivityOptionsCompat
+import androidx.core.content.ContextCompat
 import androidx.core.view.ViewCompat
+import androidx.lifecycle.Observer
+import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.codingblocks.cbonlineapp.R
 import com.codingblocks.cbonlineapp.baseclasses.BaseCBActivity
+import com.codingblocks.cbonlineapp.course.adapter.ItemClickListener
+import com.codingblocks.cbonlineapp.course.adapter.PagedCourseListAdapter
 import com.codingblocks.cbonlineapp.util.COURSE_ID
 import com.codingblocks.cbonlineapp.util.COURSE_LOGO
+import com.codingblocks.cbonlineapp.util.DividerItemDecorator
 import com.codingblocks.cbonlineapp.util.LOGO_TRANSITION_NAME
-import com.codingblocks.cbonlineapp.util.extensions.observer
-import com.codingblocks.cbonlineapp.util.extensions.setRv
 import com.codingblocks.cbonlineapp.util.extensions.setToolbar
 import kotlinx.android.synthetic.main.activity_search_course.*
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class SearchCourseActivity : BaseCBActivity() {
 
-    private val viewModel by viewModel<CourseViewModel>()
-    private val courseCardListAdapter = CourseListAdapter("LIST")
+    private val viewModel: CourseViewModel by viewModel()
+    private val courseCardListAdapter = PagedCourseListAdapter("LIST")
     val dialog: CourseSearchFragment by lazy { CourseSearchFragment() }
     private val itemClickListener: ItemClickListener by lazy {
         object : ItemClickListener {
@@ -45,18 +49,15 @@ class SearchCourseActivity : BaseCBActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_search_course)
         setToolbar(searchToolbar, title = "All Courses")
-
-        viewModel.fetchAllCourses()
-        courseSearchRv.setRv(
-            this@SearchCourseActivity,
-            courseCardListAdapter,
-            orientation = RecyclerView.VERTICAL,
-            setDivider = true
-        )
-
-        viewModel.suggestedCourses.observer(this) { courses ->
-            courseCardListAdapter.submitList(courses)
+        courseSearchRv.apply {
+            layoutManager = LinearLayoutManager(this@SearchCourseActivity, RecyclerView.VERTICAL, false)
+            addItemDecoration(DividerItemDecorator(ContextCompat.getDrawable(this@SearchCourseActivity, R.drawable.divider)!!))
+            adapter = courseCardListAdapter
         }
+
+        viewModel.getCourses().observe(this, Observer { courses ->
+            courseCardListAdapter.submitList(courses)
+        })
 
         searchBtn.setOnClickListener {
             dialog.show(supportFragmentManager, "course_search")
