+++ app/src/main/java/com/codingblocks/cbonlineapp/course/adapter/CourseDiffUtil.kt
@@ -1,0 +1,15 @@
+package com.codingblocks.cbonlineapp.course.adapter
+
+import androidx.recyclerview.widget.DiffUtil
+import com.codingblocks.cbonlineapp.util.extensions.sameAndEqual
+import com.codingblocks.onlineapi.models.Course
+
+class CourseDiffUtil : DiffUtil.ItemCallback<Course>() {
+    override fun areItemsTheSame(oldItem: Course, newItem: Course): Boolean {
+        return oldItem.id == newItem.id
+    }
+
+    override fun areContentsTheSame(oldItem: Course, newItem: Course): Boolean {
+        return oldItem.sameAndEqual(newItem)
+    }
+}
