+++ app/src/main/java/com/codingblocks/cbonlineapp/course/adapter/ItemClickListener.kt
@@ -1,0 +1,7 @@
+package com.codingblocks.cbonlineapp.course.adapter
+
+import android.widget.ImageView
+
+interface ItemClickListener {
+    fun onClick(id: String, name: String, logo: ImageView)
+}
